added diagonal a* 2d
