+++ src/com/gubertmc/plugin/main/algorithms/astar/astar2d/diagonalmovement/AstarSimulationDiagonal2D.java
@@ -1,0 +1,210 @@
+package com.gubertmc.plugin.main.algorithms.astar.astar2d.diagonalmovement;
+
+import com.gubertmc.plugin.main.algorithms.Node;
+import com.gubertmc.plugin.main.algorithms.astar.astar2d.AstarSimulation2D;
+
+import java.util.PriorityQueue;
+
+public class AstarSimulationDiagonal2D extends AstarSimulation2D {
+
+    public AstarSimulationDiagonal2D(int[][][] maze, int[] startCoordinate, int[] endCoordinate) {
+        super(maze, startCoordinate, endCoordinate);
+    }
+
+    @Override
+    public int calculateG(Node node) {
+        int row = node.getRow();
+        int col = node.getCol();
+        if (row == getCurrentNode().getRow() && col == getCurrentNode().getCol()) {
+            return 0;
+        }
+
+        Node parent = node.getParent();
+        if (parent == null) {
+            int xDistance;
+            if (col > getCurrentNode().getCol()) {
+                xDistance = col - getCurrentNode().getCol();
+            } else {
+                xDistance = getCurrentNode().getCol() - col;
+            }
+
+            int yDistance;
+            if (row > getCurrentNode().getRow()) {
+                yDistance = row - getCurrentNode().getRow();
+            } else {
+                yDistance = getCurrentNode().getRow() - row;
+            }
+
+            if (xDistance != 0 && yDistance != 0) {
+                return 14;
+            } else {
+                return (xDistance * 10) + (yDistance * 10);
+            }
+        }
+
+        if (col != parent.getCol() && row != parent.getRow()) {
+            return 14 + parent.getG();
+        }
+
+        return 10 + parent.getG();
+    }
+
+    @Override
+    public void calculateNeighborValues() {
+        int row = getCurrentNode().getRow();
+        int col = getCurrentNode().getCol();
+        int zNum = 0;
+
+        // north
+        if (row - 1 > -1 && getGrid()[row - 1][col][zNum].getType() == 0
+                && !getClosedList().contains(getGrid()[row - 1][col][zNum])) {
+            Node[][][] grid = getGrid();
+            grid[row - 1][col][zNum].setParent(getCurrentNode());
+            int g = calculateG(grid[row - 1][col][zNum]);
+            grid[row - 1][col][zNum].setG(g);
+            int h = calculateH(grid[row - 1][col][zNum]);
+            grid[row - 1][col][zNum].setH(h);
+            grid[row - 1][col][zNum].setF();
+            setGrid(grid);
+            PriorityQueue<Node> openList = getOpenList();
+            openList.add(grid[row - 1][col][zNum]);
+            setOpenList(openList);
+            int[][][] tileGrid = getTileGrid();
+            tileGrid[row - 1][col][zNum] = 2;
+            setTileGrid(tileGrid);
+        }
+
+        // northeast
+        if (row - 1 > -1 && col + 1 < getSize() && getGrid()[row - 1][col + 1][zNum].getType() == 0
+                && !getClosedList().contains(getGrid()[row - 1][col + 1][zNum])) {
+            Node[][][] grid = getGrid();
+            grid[row - 1][col + 1][zNum].setParent(getCurrentNode());
+            int g = calculateG(grid[row - 1][col + 1][zNum]);
+            grid[row - 1][col + 1][zNum].setG(g);
+            int h = calculateH(grid[row - 1][col + 1][zNum]);
+            grid[row - 1][col + 1][zNum].setH(h);
+            grid[row - 1][col + 1][zNum].setF();
+            setGrid(grid);
+            PriorityQueue<Node> openList = getOpenList();
+            openList.add(grid[row - 1][col + 1][zNum]);
+            setOpenList(openList);
+            int[][][] tileGrid = getTileGrid();
+            tileGrid[row - 1][col + 1][zNum] = 2;
+            setTileGrid(tileGrid);
+        }
+
+        // east
+        if (col + 1 < getSize() && getGrid()[row][col + 1][zNum].getType() == 0
+                && !getClosedList().contains(getGrid()[row][col + 1][zNum])) {
+            Node[][][] grid = getGrid();
+            grid[row][col + 1][zNum].setParent(getCurrentNode());
+            int g = calculateG(grid[row][col + 1][zNum]);
+            grid[row][col + 1][zNum].setG(g);
+            int h = calculateH(grid[row][col + 1][zNum]);
+            grid[row][col + 1][zNum].setH(h);
+            grid[row][col + 1][zNum].setF();
+            setGrid(grid);
+            PriorityQueue<Node> openList = getOpenList();
+            openList.add(grid[row][col + 1][zNum]);
+            setOpenList(openList);
+            int[][][] tileGrid = getTileGrid();
+            tileGrid[row][col + 1][zNum] = 2;
+            setTileGrid(tileGrid);
+        }
+
+        // southeast
+        if (row + 1 < getSize() && col + 1 < getSize() && getGrid()[row + 1][col + 1][zNum].getType() == 0
+                && !getClosedList().contains(getGrid()[row + 1][col + 1][zNum])) {
+            Node[][][] grid = getGrid();
+            grid[row + 1][col + 1][zNum].setParent(getCurrentNode());
+            int g = calculateG(grid[row + 1][col + 1][zNum]);
+            grid[row + 1][col + 1][zNum].setG(g);
+            int h = calculateH(grid[row + 1][col + 1][zNum]);
+            grid[row + 1][col + 1][zNum].setH(h);
+            grid[row + 1][col + 1][zNum].setF();
+            setGrid(grid);
+            PriorityQueue<Node> openList = getOpenList();
+            openList.add(grid[row + 1][col + 1][zNum]);
+            setOpenList(openList);
+            int[][][] tileGrid = getTileGrid();
+            tileGrid[row + 1][col + 1][zNum] = 2;
+            setTileGrid(tileGrid);
+        }
+
+        // south
+        if (row + 1 < getSize() && getGrid()[row + 1][col][zNum].getType() == 0
+                && !getClosedList().contains(getGrid()[row + 1][col][zNum])) {
+            Node[][][] grid = getGrid();
+            grid[row + 1][col][zNum].setParent(getCurrentNode());
+            int g = calculateG(grid[row + 1][col][zNum]);
+            grid[row + 1][col][zNum].setG(g);
+            int h = calculateH(grid[row + 1][col][zNum]);
+            grid[row + 1][col][zNum].setH(h);
+            grid[row + 1][col][zNum].setF();
+            setGrid(grid);
+            PriorityQueue<Node> openList = getOpenList();
+            openList.add(grid[row + 1][col][zNum]);
+            setOpenList(openList);
+            int[][][] tileGrid = getTileGrid();
+            tileGrid[row + 1][col][zNum] = 2;
+            setTileGrid(tileGrid);
+        }
+
+        // southwest
+        if (row + 1 < getSize() && col - 1 > -1 && getGrid()[row + 1][col - 1][zNum].getType() == 0
+                && !getClosedList().contains(getGrid()[row + 1][col - 1][zNum])) {
+            Node[][][] grid = getGrid();
+            grid[row + 1][col - 1][zNum].setParent(getCurrentNode());
+            int g = calculateG(grid[row + 1][col - 1][zNum]);
+            grid[row + 1][col - 1][zNum].setG(g);
+            int h = calculateH(grid[row + 1][col - 1][zNum]);
+            grid[row + 1][col - 1][zNum].setH(h);
+            grid[row + 1][col - 1][zNum].setF();
+            setGrid(grid);
+            PriorityQueue<Node> openList = getOpenList();
+            openList.add(grid[row + 1][col - 1][zNum]);
+            setOpenList(openList);
+            int[][][] tileGrid = getTileGrid();
+            tileGrid[row + 1][col - 1][zNum] = 2;
+            setTileGrid(tileGrid);
+        }
+
+        // west
+        if (col - 1 > -1 && getGrid()[row][col - 1][zNum].getType() == 0
+                && !getClosedList().contains(getGrid()[row][col - 1][zNum])) {
+            Node[][][] grid = getGrid();
+            grid[row][col - 1][zNum].setParent(getCurrentNode());
+            int g = calculateG(grid[row][col - 1][zNum]);
+            grid[row][col - 1][zNum].setG(g);
+            int h = calculateH(grid[row][col - 1][zNum]);
+            grid[row][col - 1][zNum].setH(h);
+            grid[row][col - 1][zNum].setF();
+            setGrid(grid);
+            PriorityQueue<Node> openList = getOpenList();
+            openList.add(grid[row][col - 1][zNum]);
+            setOpenList(openList);
+            int[][][] tileGrid = getTileGrid();
+            tileGrid[row][col - 1][zNum] = 2;
+            setTileGrid(tileGrid);
+        }
+
+        // northwest
+        if (row - 1 > -1 && col - 1 > -1 && getGrid()[row - 1][col - 1][zNum].getType() == 0
+                && !getClosedList().contains(getGrid()[row - 1][col - 1][zNum])) {
+            Node[][][] grid = getGrid();
+            grid[row - 1][col - 1][zNum].setParent(getCurrentNode());
+            int g = calculateG(grid[row - 1][col - 1][zNum]);
+            grid[row - 1][col - 1][zNum].setG(g);
+            int h = calculateH(grid[row - 1][col - 1][zNum]);
+            grid[row - 1][col - 1][zNum].setH(h);
+            grid[row - 1][col - 1][zNum].setF();
+            setGrid(grid);
+            PriorityQueue<Node> openList = getOpenList();
+            openList.add(grid[row - 1][col - 1][zNum]);
+            setOpenList(openList);
+            int[][][] tileGrid = getTileGrid();
+            tileGrid[row - 1][col - 1][zNum] = 2;
+            setTileGrid(tileGrid);
+        }
+    }
+}
