+++ src/com/gubertmc/plugin/main/algorithms/astar/astar3d/nondiagonalmovement/AstarAnimation3D.java
@@ -1,4 +1,4 @@
-package com.gubertmc.plugin.main.algorithms.greedybestfirstsearch.greedybestfirstsearch2d;
+package com.gubertmc.plugin.main.algorithms.astar.astar3d.nondiagonalmovement;
 
 import com.gubertmc.MazeGeneratorPlugin;
 import com.gubertmc.plugin.main.algorithms.Animation;
@@ -10,9 +10,9 @@ import org.bukkit.scheduler.BukkitRunnable;
 import java.util.ArrayList;
 import java.util.PriorityQueue;
 
-public class GreedyBestFirstSearchAnimation2D extends Animation {
+public class AstarAnimation3D extends Animation {
 
-    public GreedyBestFirstSearchAnimation2D(
+    public AstarAnimation3D(
             MazeGeneratorPlugin plugin,
             Location[][][] tiles,
             int[] startCoordinate,
@@ -37,7 +37,7 @@ public class GreedyBestFirstSearchAnimation2D extends Animation {
                 groundMaterial,
                 startGlassMaterial,
                 endGlassMaterial,
-                false
+                true
         );
     }
 
@@ -55,33 +55,37 @@ public class GreedyBestFirstSearchAnimation2D extends Animation {
         }
         setTileGridInt(tempArray);
 
-        setCurrentNode(new Node(getStartCoordinate()[1], getStartCoordinate()[0], -1, 0));
-        setEndNode(new Node(getEndCoordinate()[1], getEndCoordinate()[0], -1, 0));
+        setCurrentNode(new Node(getStartCoordinate()[1], getStartCoordinate()[0], getStartCoordinate()[2], 0));
+        setEndNode(new Node(getEndCoordinate()[1], getEndCoordinate()[0], getEndCoordinate()[2], 0));
         Node[][][] grid = getGrid();
-        grid[getStartCoordinate()[1]][getStartCoordinate()[0]][0] = getCurrentNode();
+        grid[getStartCoordinate()[1]][getStartCoordinate()[0]][getStartCoordinate()[2]] = getCurrentNode();
         setGrid(grid);
-        grid[getEndCoordinate()[1]][getEndCoordinate()[0]][0] = getEndNode();
+        grid[getEndCoordinate()[1]][getEndCoordinate()[0]][getEndCoordinate()[2]] = getEndNode();
         setGrid(grid);
 
         for (int i = 0; i < getSize(); i++) {
             for (int j = 0; j < getSize(); j++) {
-                if (getTileGrid()[i][j][0].getBlock().getType() == Material.AIR) {
-                    Node node = new Node(i, j, -1, 0);
-                    grid[i][j][0] = node;
-                    setGrid(grid);
-                }
-                if (getTileGrid()[i][j][0].getBlock().getType() == getWallMaterial()) {
-                    Node node = new Node(i, j, -1, 1);
-                    grid[i][j][0] = node;
-                    setGrid(grid);
+                for (int k = 0; k < getSize(); k++) {
+                    if (getTileGrid()[i][j][k].getBlock().getType() == getPathGroundMaterial()) {
+                        Node node = new Node(i, j, k, 0);
+                        grid[i][j][k] = node;
+                        setGrid(grid);
+                    }
+                    if (getTileGrid()[i][j][k].getBlock().getType() == getWallMaterial()) {
+                        Node node = new Node(i, j, k, 1);
+                        grid[i][j][k] = node;
+                        setGrid(grid);
+                    }
                 }
             }
         }
 
         Node currentNode = getCurrentNode();
+        int g = calculateG(currentNode);
+        currentNode.setG(g);
         int h = calculateH(currentNode);
         currentNode.setH(h);
-        currentNode.setBfsF();
+        currentNode.setF();
         setCurrentNode(currentNode);
         setStartNode(currentNode);
         PriorityQueue<Node> openList = getOpenList();
@@ -108,11 +112,11 @@ public class GreedyBestFirstSearchAnimation2D extends Animation {
                 for (int i = path.size() - 1; i > -1; i--) {
                     int row = path.get(i).getRow();
                     int col = path.get(i).getCol();
-                    int zNum = 0;
+                    int zNum = path.get(i).getZ();
 
                     if (getTileGridInt()[row][col][zNum] == 1) {
                         int x = getTileGrid()[row][col][zNum].getBlockX();
-                        int y = getTileGrid()[row][col][0].getBlockY() - 1;
+                        int y = getTileGrid()[row][col][zNum].getBlockY();
                         int z = getTileGrid()[row][col][zNum].getBlockZ();
 
                         Location floor = new Location(getTileGrid()[row][col][zNum].getWorld(), x, y, z);
@@ -128,6 +132,7 @@ public class GreedyBestFirstSearchAnimation2D extends Animation {
                 } catch (NullPointerException e) {
                     System.out.println(e);
                 }
+
                 try {
                     assert getOpenList().peek() != null;
                 } catch (NullPointerException e) {
@@ -173,7 +178,7 @@ public class GreedyBestFirstSearchAnimation2D extends Animation {
         new BukkitRunnable() {
             @Override
             public void run() {
-                if (loc.getBlock().getType() != getEndPointGlass()) {
+                if (loc.getBlock().getType() != Material.BEACON) {
                     loc.getBlock().setType(material);
                 }
                 cancel();
@@ -181,11 +186,53 @@ public class GreedyBestFirstSearchAnimation2D extends Animation {
         }.runTaskTimer(getPlugin(), time, 20L);
     }
 
+    public int calculateG(Node node) {
+        int row = node.getRow();
+        int col = node.getCol();
+        int zNum = node.getZ();
+        if (row == getCurrentNode().getRow() && col == getCurrentNode().getCol() && zNum == getCurrentNode().getZ()) {
+            return 0;
+        }
+
+        Node parent = node.getParent();
+        if (parent == null) {
+            int xDistance;
+            if (col > getCurrentNode().getCol()) {
+                xDistance = col - getCurrentNode().getCol();
+            } else {
+                xDistance = getCurrentNode().getCol() - col;
+            }
+
+            int yDistance;
+            if (row > getCurrentNode().getRow()) {
+                yDistance = row - getCurrentNode().getRow();
+            } else {
+                yDistance = getCurrentNode().getRow() - row;
+            }
+
+            int zDistance;
+            if (zNum > getCurrentNode().getZ()) {
+                zDistance = zNum - getCurrentNode().getZ();
+            } else {
+                zDistance = getCurrentNode().getZ() - zNum;
+            }
+
+            if (zNum == -1) {
+                zDistance = 0;
+            }
+
+            return (xDistance * 10) + (yDistance * 10) + (zDistance * 10);
+        }
+        return 10 + parent.getG();
+    }
+
     public int calculateH(Node node) {
         int row = node.getRow();
         int col = node.getCol();
+        int zNum = node.getZ();
         int x = 0;
         int y = 0;
+        int z = 0;
 
         while (col < getEndNode().getCol() || col > getEndNode().getCol()) {
             x += 10;
@@ -205,8 +252,17 @@ public class GreedyBestFirstSearchAnimation2D extends Animation {
                 row--;
             }
         }
+        while (zNum < getEndNode().getZ() || zNum > getEndNode().getZ()) {
+            z += 10;
+            if (zNum < getEndNode().getZ()) {
+                zNum++;
+            }
+            if (zNum > getEndNode().getZ()) {
+                zNum--;
+            }
+        }
 
-        return x + y;
+        return x + y + z;
     }
 
     public ArrayList<Node> generatePath() {
@@ -223,16 +279,18 @@ public class GreedyBestFirstSearchAnimation2D extends Animation {
     public void calculateNeighborValues() {
         int row = getCurrentNode().getRow();
         int col = getCurrentNode().getCol();
-        int zNum = 0;
+        int zNum = getCurrentNode().getZ();
 
         // front node
         if (row - 1 > -1 && getGrid()[row - 1][col][zNum].getType() == 0
                 && !getClosedList().contains(getGrid()[row - 1][col][zNum])) {
             Node[][][] grid = getGrid();
             grid[row - 1][col][zNum].setParent(getCurrentNode());
+            int g = calculateG(grid[row - 1][col][zNum]);
+            grid[row - 1][col][zNum].setG(g);
             int h = calculateH(grid[row - 1][col][zNum]);
             grid[row - 1][col][zNum].setH(h);
-            grid[row - 1][col][zNum].setBfsF();
+            grid[row - 1][col][zNum].setF();
             setGrid(grid);
             PriorityQueue<Node> openList = getOpenList();
             openList.add(grid[row - 1][col][zNum]);
@@ -244,7 +302,7 @@ public class GreedyBestFirstSearchAnimation2D extends Animation {
             Location loc = getTileGrid()[row - 1][col][zNum];
             if (!getExploredPlaces().contains(loc)) {
                 loc = new Location(
-                        loc.getWorld(), loc.getBlock().getX(), loc.getBlock().getY() - 1, loc.getBlock().getZ()
+                        loc.getWorld(), loc.getBlock().getX(), loc.getBlock().getY(), loc.getBlock().getZ()
                 );
                 ArrayList<Location> exploredPlaces = getExploredPlaces();
                 exploredPlaces.add(loc);
@@ -257,9 +315,11 @@ public class GreedyBestFirstSearchAnimation2D extends Animation {
                 && !getClosedList().contains(getGrid()[row][col + 1][zNum])) {
             Node[][][] grid = getGrid();
             grid[row][col + 1][zNum].setParent(getCurrentNode());
+            int g = calculateG(grid[row][col + 1][zNum]);
+            grid[row][col + 1][zNum].setG(g);
             int h = calculateH(grid[row][col + 1][zNum]);
             grid[row][col + 1][zNum].setH(h);
-            grid[row][col + 1][zNum].setBfsF();
+            grid[row][col + 1][zNum].setF();
             setGrid(grid);
             PriorityQueue<Node> openList = getOpenList();
             openList.add(grid[row][col + 1][zNum]);
@@ -271,7 +331,7 @@ public class GreedyBestFirstSearchAnimation2D extends Animation {
             Location loc = getTileGrid()[row][col + 1][zNum];
             if (!getExploredPlaces().contains(loc)) {
                 loc = new Location(
-                        loc.getWorld(), loc.getBlock().getX(), loc.getBlock().getY() - 1, loc.getBlock().getZ()
+                        loc.getWorld(), loc.getBlock().getX(), loc.getBlock().getY(), loc.getBlock().getZ()
                 );
                 ArrayList<Location> exploredPlaces = getExploredPlaces();
                 exploredPlaces.add(loc);
@@ -284,9 +344,11 @@ public class GreedyBestFirstSearchAnimation2D extends Animation {
                 && !getClosedList().contains(getGrid()[row + 1][col][zNum])) {
             Node[][][] grid = getGrid();
             grid[row + 1][col][zNum].setParent(getCurrentNode());
+            int g = calculateG(grid[row + 1][col][zNum]);
+            grid[row + 1][col][zNum].setG(g);
             int h = calculateH(grid[row + 1][col][zNum]);
             grid[row + 1][col][zNum].setH(h);
-            grid[row + 1][col][zNum].setBfsF();
+            grid[row + 1][col][zNum].setF();
             setGrid(grid);
             PriorityQueue<Node> openList = getOpenList();
             openList.add(grid[row + 1][col][zNum]);
@@ -298,7 +360,7 @@ public class GreedyBestFirstSearchAnimation2D extends Animation {
             Location loc = getTileGrid()[row + 1][col][zNum];
             if (!getExploredPlaces().contains(loc)) {
                 loc = new Location(
-                        loc.getWorld(), loc.getBlock().getX(), loc.getBlock().getY() - 1, loc.getBlock().getZ()
+                        loc.getWorld(), loc.getBlock().getX(), loc.getBlock().getY(), loc.getBlock().getZ()
                 );
                 ArrayList<Location> exploredPlaces = getExploredPlaces();
                 exploredPlaces.add(loc);
@@ -311,9 +373,11 @@ public class GreedyBestFirstSearchAnimation2D extends Animation {
                 && !getClosedList().contains(getGrid()[row][col - 1][zNum])) {
             Node[][][] grid = getGrid();
             grid[row][col - 1][zNum].setParent(getCurrentNode());
+            int g = calculateG(grid[row][col - 1][zNum]);
+            grid[row][col - 1][zNum].setG(g);
             int h = calculateH(grid[row][col - 1][zNum]);
             grid[row][col - 1][zNum].setH(h);
-            grid[row][col - 1][zNum].setBfsF();
+            grid[row][col - 1][zNum].setF();
             setGrid(grid);
             PriorityQueue<Node> openList = getOpenList();
             openList.add(grid[row][col - 1][zNum]);
@@ -325,7 +389,63 @@ public class GreedyBestFirstSearchAnimation2D extends Animation {
             Location loc = getTileGrid()[row][col - 1][zNum];
             if (!getExploredPlaces().contains(loc)) {
                 loc = new Location(
-                        loc.getWorld(), loc.getBlock().getX(), loc.getBlock().getY() - 1, loc.getBlock().getZ()
+                        loc.getWorld(), loc.getBlock().getX(), loc.getBlock().getY(), loc.getBlock().getZ()
+                );
+                ArrayList<Location> exploredPlaces = getExploredPlaces();
+                exploredPlaces.add(loc);
+                setExploredPlaces(exploredPlaces);
+            }
+        }
+
+        // bottom node
+        if (zNum - 1 > -1 && getGrid()[row][col][zNum - 1].getType() == 0
+                && !getClosedList().contains(getGrid()[row][col][zNum - 1])) {
+            Node[][][] grid = getGrid();
+            grid[row][col][zNum - 1].setParent(getCurrentNode());
+            int g = calculateG(grid[row][col][zNum - 1]);
+            grid[row][col][zNum - 1].setG(g);
+            int h = calculateH(grid[row][col][zNum - 1]);
+            grid[row][col][zNum - 1].setH(h);
+            grid[row][col][zNum - 1].setF();
+            setGrid(grid);
+            PriorityQueue<Node> openList = getOpenList();
+            openList.add(grid[row][col][zNum - 1]);
+            int[][][] tileGridInt = getTileGridInt();
+            tileGridInt[row][col][zNum - 1] = 1;
+            setTileGridInt(tileGridInt);
+
+            Location loc = getTileGrid()[row][col][zNum - 1];
+            if (!getExploredPlaces().contains(loc)) {
+                loc = new Location(
+                        loc.getWorld(), loc.getBlock().getX(), loc.getBlock().getY(), loc.getBlock().getZ()
+                );
+                ArrayList<Location> exploredPlaces = getExploredPlaces();
+                exploredPlaces.add(loc);
+                setExploredPlaces(exploredPlaces);
+            }
+        }
+
+        // top node
+        if (zNum + 1 < getSize() && getGrid()[row][col][zNum + 1].getType() == 0
+                && !getClosedList().contains(getGrid()[row][col][zNum + 1])) {
+            Node[][][] grid = getGrid();
+            grid[row][col][zNum + 1].setParent(getCurrentNode());
+            int g = calculateG(grid[row][col][zNum + 1]);
+            grid[row][col][zNum + 1].setG(g);
+            int h = calculateH(grid[row][col][zNum + 1]);
+            grid[row][col][zNum + 1].setH(h);
+            grid[row][col][zNum + 1].setF();
+            setGrid(grid);
+            PriorityQueue<Node> openList = getOpenList();
+            openList.add(grid[row][col][zNum + 1]);
+            int[][][] tileGridInt = getTileGridInt();
+            tileGridInt[row][col][zNum + 1] = 1;
+            setTileGridInt(tileGridInt);
+
+            Location loc = getTileGrid()[row][col][zNum + 1];
+            if (!getExploredPlaces().contains(loc)) {
+                loc = new Location(
+                        loc.getWorld(), loc.getBlock().getX(), loc.getBlock().getY(), loc.getBlock().getZ()
                 );
                 ArrayList<Location> exploredPlaces = getExploredPlaces();
                 exploredPlaces.add(loc);
