+++ src/com/gubertmc/plugin/main/algorithms/greedybestfirstsearch/greedybestfirstsearch2d/nondiagonalmovement/GreedyBestFirstSearchSimulation2D.java
@@ -1,48 +1,45 @@
-package com.gubertmc.plugin.main.algorithms.astar.astar3d;
+package com.gubertmc.plugin.main.algorithms.greedybestfirstsearch.greedybestfirstsearch2d.nondiagonalmovement;
 
-import com.gubertmc.plugin.main.algorithms.Simulation;
 import com.gubertmc.plugin.main.algorithms.Node;
+import com.gubertmc.plugin.main.algorithms.Simulation;
 
 import java.util.ArrayList;
 import java.util.PriorityQueue;
 
-public class AstarSimulation3D extends Simulation {
+public class GreedyBestFirstSearchSimulation2D extends Simulation {
 
-    public AstarSimulation3D(int[][][] maze, int[] startCoordinate, int[] endCoordinate) {
-        super(maze, startCoordinate, endCoordinate, true);
+    public GreedyBestFirstSearchSimulation2D(int[][][] maze, int[] startCoordinate, int[] endCoordinate) {
+        super(maze, startCoordinate, endCoordinate, false);
     }
 
     @Override
     public void setup() {
-        setCurrentNode(new Node(getStartCoordinate()[1], getStartCoordinate()[0], getStartCoordinate()[2], 0));
-        setEndNode(new Node(getEndCoordinate()[1], getEndCoordinate()[0], getEndCoordinate()[2], 0));
+        setCurrentNode(new Node(getStartCoordinate()[1], getStartCoordinate()[0], -1, 0));
+        setEndNode(new Node(getEndCoordinate()[1], getEndCoordinate()[0], -1, 0));
         Node[][][] grid = getGrid();
-        grid[getStartCoordinate()[1]][getStartCoordinate()[0]][getStartCoordinate()[2]] = getCurrentNode();
-        grid[getEndCoordinate()[1]][getEndCoordinate()[0]][getEndCoordinate()[2]] = getEndNode();
+        grid[getStartCoordinate()[1]][getStartCoordinate()[0]][0] = getCurrentNode();
+        grid[getEndCoordinate()[1]][getEndCoordinate()[0]][0] = getEndNode();
+        setGrid(grid);
 
         for (int i = 0; i < getSize(); i++) {
             for (int j = 0; j < getSize(); j++) {
-                for (int k = 0; k < getSize(); k++) {
-                    if (getTileGrid()[i][j][k] == 0) {
-                        Node node = new Node(i, j, k, 0);
-                        grid[i][j][k] = node;
-                        setGrid(grid);
-                    }
-                    if (getTileGrid()[i][j][k] == 1) {
-                        Node node = new Node(i, j, k, 1);
-                        grid[i][j][k] = node;
-                        setGrid(grid);
-                    }
+                if (getTileGrid()[i][j][0] == 0) {
+                    Node node = new Node(i, j, -1, 0);
+                    grid[i][j][0] = node;
+                    setGrid(grid);
+                }
+                if (getTileGrid()[i][j][0] == 1) {
+                    Node node = new Node(i, j, -1, 1);
+                    grid[i][j][0] = node;
+                    setGrid(grid);
                 }
             }
         }
 
         Node currentNode = getCurrentNode();
-        int g = calculateG(currentNode);
-        currentNode.setG(g);
         int h = calculateH(currentNode);
         currentNode.setH(h);
-        currentNode.setF();
+        currentNode.setBfsF();
         setCurrentNode(currentNode);
         setStartNode(currentNode);
         PriorityQueue<Node> openList = getOpenList();
@@ -68,7 +65,8 @@ public class AstarSimulation3D extends Simulation {
                 for (int i = path.size() - 1; i > -1; i--) {
                     int row = path.get(i).getRow();
                     int col = path.get(i).getCol();
-                    int zNum = path.get(i).getZ();
+                    int zNum = 0;
+
                     if (getTileGrid()[row][col][zNum] == 2) {
                         int[][][] tileGrid = getTileGrid();
                         tileGrid[row][col][zNum] = 3;
@@ -83,8 +81,8 @@ public class AstarSimulation3D extends Simulation {
                     System.out.println(e.getMessage());
                 }
                 int[][][] tileGrid = getTileGrid();
-                tileGrid[getStartNode().getRow()][getStartNode().getCol()][getStartNode().getZ()] = 4;
-                tileGrid[getEndNode().getRow()][getEndNode().getCol()][getEndNode().getZ()] = 5;
+                tileGrid[getStartNode().getRow()][getStartNode().getCol()][0] = 4;
+                tileGrid[getEndNode().getRow()][getEndNode().getCol()][0] = 5;
                 setTileGrid(tileGrid);
                 try {
                     assert getOpenList().peek() != null;
@@ -103,49 +101,11 @@ public class AstarSimulation3D extends Simulation {
         return pathFound;
     }
 
-    public int calculateG(Node node) {
-        int row = node.getRow();
-        int col = node.getCol();
-        int zNum = node.getZ();
-        if (row == getCurrentNode().getRow() && col == getCurrentNode().getCol() && zNum == getCurrentNode().getZ()) {
-            return 0;
-        }
-
-        Node parent = node.getParent();
-        if (parent == null) {
-            int xDistance;
-            if (col > getCurrentNode().getCol()) {
-                xDistance = col - getCurrentNode().getCol();
-            } else {
-                xDistance = getCurrentNode().getCol() - col;
-            }
-
-            int yDistance;
-            if (row > getCurrentNode().getRow()) {
-                yDistance = row - getCurrentNode().getRow();
-            } else {
-                yDistance = getCurrentNode().getRow() - row;
-            }
-
-            int zDistance;
-            if (zNum > getCurrentNode().getZ()) {
-                zDistance = zNum - getCurrentNode().getZ();
-            } else {
-                zDistance = getCurrentNode().getZ() - zNum;
-            }
-
-            return (xDistance * 10) + (yDistance * 10) + (zDistance * 10);
-        }
-        return 10 + parent.getG();
-    }
-
     public int calculateH(Node node) {
         int row = node.getRow();
         int col = node.getCol();
-        int zNum = node.getZ();
         int x = 0;
         int y = 0;
-        int z = 0;
 
         while (col < getEndNode().getCol() || col > getEndNode().getCol()) {
             x += 10;
@@ -165,34 +125,23 @@ public class AstarSimulation3D extends Simulation {
                 row--;
             }
         }
-        while (zNum < getEndNode().getZ() || zNum > getEndNode().getZ()) {
-            z += 10;
-            if (zNum < getEndNode().getZ()) {
-                zNum++;
-            }
-            if (zNum > getEndNode().getZ()) {
-                zNum--;
-            }
-        }
 
-        return x + y + z;
+        return x + y;
     }
 
     public void calculateNeighborValues() {
         int row = getCurrentNode().getRow();
         int col = getCurrentNode().getCol();
-        int zNum = getCurrentNode().getZ();
+        int zNum = 0;
 
         // front node
         if (row - 1 > -1 && getGrid()[row - 1][col][zNum].getType() == 0
                 && !getClosedList().contains(getGrid()[row - 1][col][zNum])) {
             Node[][][] grid = getGrid();
             grid[row - 1][col][zNum].setParent(getCurrentNode());
-            int g = calculateG(grid[row - 1][col][zNum]);
-            grid[row - 1][col][zNum].setG(g);
             int h = calculateH(grid[row - 1][col][zNum]);
             grid[row - 1][col][zNum].setH(h);
-            grid[row - 1][col][zNum].setF();
+            grid[row - 1][col][zNum].setBfsF();
             setGrid(grid);
             PriorityQueue<Node> openList = getOpenList();
             openList.add(grid[row - 1][col][zNum]);
@@ -207,11 +156,9 @@ public class AstarSimulation3D extends Simulation {
                 && !getClosedList().contains(getGrid()[row][col + 1][zNum])) {
             Node[][][] grid = getGrid();
             grid[row][col + 1][zNum].setParent(getCurrentNode());
-            int g = calculateG(grid[row][col + 1][zNum]);
-            grid[row][col + 1][zNum].setG(g);
             int h = calculateH(grid[row][col + 1][zNum]);
             grid[row][col + 1][zNum].setH(h);
-            grid[row][col + 1][zNum].setF();
+            grid[row][col + 1][zNum].setBfsF();
             setGrid(grid);
             PriorityQueue<Node> openList = getOpenList();
             openList.add(grid[row][col + 1][zNum]);
@@ -226,11 +173,9 @@ public class AstarSimulation3D extends Simulation {
                 && !getClosedList().contains(getGrid()[row + 1][col][zNum])) {
             Node[][][] grid = getGrid();
             grid[row + 1][col][zNum].setParent(getCurrentNode());
-            int g = calculateG(grid[row + 1][col][zNum]);
-            grid[row + 1][col][zNum].setG(g);
             int h = calculateH(grid[row + 1][col][zNum]);
             grid[row + 1][col][zNum].setH(h);
-            grid[row + 1][col][zNum].setF();
+            grid[row + 1][col][zNum].setBfsF();
             setGrid(grid);
             PriorityQueue<Node> openList = getOpenList();
             openList.add(grid[row + 1][col][zNum]);
@@ -245,55 +190,15 @@ public class AstarSimulation3D extends Simulation {
                 && !getClosedList().contains(getGrid()[row][col - 1][zNum])) {
             Node[][][] grid = getGrid();
             grid[row][col - 1][zNum].setParent(getCurrentNode());
-            int g = calculateG(grid[row][col - 1][zNum]);
-            grid[row][col - 1][zNum].setG(g);
             int h = calculateH(grid[row][col - 1][zNum]);
             grid[row][col - 1][zNum].setH(h);
-            grid[row][col - 1][zNum].setF();
+            grid[row][col - 1][zNum].setBfsF();
             setGrid(grid);
             PriorityQueue<Node> openList = getOpenList();
             openList.add(grid[row][col - 1][zNum]);
             setOpenList(openList);
             int[][][] tileGrid = getTileGrid();
             tileGrid[row][col - 1][zNum] = 2;
-            setTileGrid(tileGrid);
-        }
-
-        // bottom node
-        if (zNum - 1 > -1 && getGrid()[row][col][zNum - 1].getType() == 0
-                && !getClosedList().contains(getGrid()[row][col][zNum - 1])) {
-            Node[][][] grid = getGrid();
-            grid[row][col][zNum - 1].setParent(getCurrentNode());
-            int g = calculateG(grid[row][col][zNum - 1]);
-            grid[row][col][zNum - 1].setG(g);
-            int h = calculateH(grid[row][col][zNum - 1]);
-            grid[row][col][zNum - 1].setH(h);
-            grid[row][col][zNum - 1].setF();
-            setGrid(grid);
-            PriorityQueue<Node> openList = getOpenList();
-            openList.add(grid[row][col][zNum - 1]);
-            setOpenList(openList);
-            int[][][] tileGrid = getTileGrid();
-            tileGrid[row][col][zNum - 1] = 2;
-            setTileGrid(tileGrid);
-        }
-
-        // top node
-        if (zNum + 1 < getSize() && getGrid()[row][col][zNum + 1].getType() == 0
-                && !getClosedList().contains(getGrid()[row][col][zNum + 1])) {
-            Node[][][] grid = getGrid();
-            grid[row][col][zNum + 1].setParent(getCurrentNode());
-            int g = calculateG(grid[row][col][zNum + 1]);
-            grid[row][col][zNum + 1].setG(g);
-            int h = calculateH(grid[row][col][zNum + 1]);
-            grid[row][col][zNum + 1].setH(h);
-            grid[row][col][zNum + 1].setF();
-            setGrid(grid);
-            PriorityQueue<Node> openList = getOpenList();
-            openList.add(grid[row][col][zNum + 1]);
-            setOpenList(openList);
-            int[][][] tileGrid = getTileGrid();
-            tileGrid[row][col][zNum + 1] = 2;
             setTileGrid(tileGrid);
         }
     }
