+++ src/com/gubertmc/plugin/main/mazes/custom/AstarMaze3D.java
@@ -3,8 +3,8 @@ package com.gubertmc.plugin.main.mazes.custom;
 import com.gubertmc.MazeGeneratorPlugin;
 import com.gubertmc.plugin.main.algorithms.Animation;
 import com.gubertmc.plugin.main.algorithms.Simulation;
-import com.gubertmc.plugin.main.algorithms.astar.astar3d.AstarAnimation3D;
-import com.gubertmc.plugin.main.algorithms.astar.astar3d.AstarSimulation3D;
+import com.gubertmc.plugin.main.algorithms.astar.astar3d.nondiagonalmovement.AstarAnimation3D;
+import com.gubertmc.plugin.main.algorithms.astar.astar3d.nondiagonalmovement.AstarSimulation3D;
 import com.gubertmc.plugin.main.mazes.Maze3D;
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
