+++ src/com/gubertmc/plugin/main/mazes/custom/AstarMazeDiagonal2D.java
@@ -3,8 +3,8 @@ package com.gubertmc.plugin.main.mazes.custom;
 import com.gubertmc.MazeGeneratorPlugin;
 import com.gubertmc.plugin.main.algorithms.Animation;
 import com.gubertmc.plugin.main.algorithms.Simulation;
-import com.gubertmc.plugin.main.algorithms.greedybestfirstsearch.greedybestfirstsearch2d.nondiagonalmovement.GreedyBestFirstSearchAnimation2D;
-import com.gubertmc.plugin.main.algorithms.greedybestfirstsearch.greedybestfirstsearch2d.nondiagonalmovement.GreedyBestFirstSearchSimulation2D;
+import com.gubertmc.plugin.main.algorithms.astar.astar2d.diagonalmovement.AstarAnimationDiagonal2D;
+import com.gubertmc.plugin.main.algorithms.astar.astar2d.diagonalmovement.AstarSimulationDiagonal2D;
 import com.gubertmc.plugin.main.mazes.Maze2D;
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
@@ -13,9 +13,9 @@ import org.bukkit.scheduler.BukkitRunnable;
 
 import static org.bukkit.Bukkit.getServer;
 
-public class GreedyBestFirstSearchMaze2D extends Maze2D {
+public class AstarMazeDiagonal2D extends Maze2D {
 
-    public GreedyBestFirstSearchMaze2D(MazeGeneratorPlugin plugin, Block block, int size, double wallPercentage) {
+    public AstarMazeDiagonal2D(MazeGeneratorPlugin plugin, Block block, int size, double wallPercentage) {
         super(plugin, block, size, wallPercentage);
     }
 
@@ -43,7 +43,7 @@ public class GreedyBestFirstSearchMaze2D extends Maze2D {
         int count = 0;
         while (!isValid()) {
             int[][][] simulationMaze = generateSimulation();
-            Simulation simulation = new GreedyBestFirstSearchSimulation2D(
+            Simulation simulation = new AstarSimulationDiagonal2D(
                     simulationMaze, getStartCoordinate(), getEndCoordinate()
             );
             simulation.setup();
@@ -63,7 +63,7 @@ public class GreedyBestFirstSearchMaze2D extends Maze2D {
                 new BukkitRunnable() {
                     @Override
                     public void run() {
-                        Animation animation = new GreedyBestFirstSearchAnimation2D(
+                        Animation animation = new AstarAnimationDiagonal2D(
                                 getPlugin(),
                                 getMazeBlockLocations(),
                                 getStartCoordinate(),
@@ -78,8 +78,8 @@ public class GreedyBestFirstSearchMaze2D extends Maze2D {
                         );
                         animation.setup();
                         setValid(animation.start());
-                        getServer().broadcastMessage(
-                                ChatColor.GREEN + "" + getSize() + "x" + getSize() + " maze generated..."
+                        getServer().broadcastMessage(ChatColor.GREEN + "" + getSize() + "x" + getSize()
+                                + " maze generated..."
                         );
                         animation.showAnimation(getTime());
                         cancel();
@@ -92,10 +92,9 @@ public class GreedyBestFirstSearchMaze2D extends Maze2D {
                 System.out.println("Invalid maze - starting new simulation...");
             }
             if (count == 50) {
-                getServer().broadcastMessage(
-                        ChatColor.RED + "A maze could not be successfully generated within 50 simulations. You " +
-                                "may experience server lag. Creating a larger maze with a lower percentage of " +
-                                "walls/blockers will greatly help and put less stress on the server."
+                getServer().broadcastMessage(ChatColor.RED + "A maze could not be successfully generated within " +
+                        "50 simulations. You may experience server lag. Creating a larger maze with a lower " +
+                        "percentage of walls/blockers will greatly help and put less stress on the server."
                 );
             }
         }
